Debugger: Do not show synthetic '_field' members for decompiled declarations
Until Kotlin Metadata contains information about backing fields, we can't know for sure if a deserialized property has a backing field.
Right now Kotlinc says it has, but this behavior brings more harm than good for the debugger.
So, as a work-around, this commit disables `_field`s for these cases.